+++ packages/http/src/mocker/negotiator/NegotiatorHelpers.ts
@@ -1,12 +1,12 @@
 import { IHttpOperation, IHttpOperationResponse, IMediaTypeContent, IHttpHeaderParam } from '@stoplight/types';
 import * as E from 'fp-ts/lib/Either';
-import { NonEmptyArray } from 'fp-ts/lib/NonEmptyArray';
-import { findIndex } from 'fp-ts/lib/Array';
+import { NonEmptyArray, fromArray } from 'fp-ts/lib/NonEmptyArray';
+import { isNonEmpty, findIndex } from 'fp-ts/lib/Array';
 import * as O from 'fp-ts/lib/Option';
-import { pipe } from 'fp-ts/lib/pipeable';
 import * as R from 'fp-ts/lib/Reader';
 import * as RE from 'fp-ts/lib/ReaderEither';
-import { get, tail } from 'lodash';
+import { pipe } from 'fp-ts/lib/pipeable';
+import { tail } from 'lodash';
 import { Logger } from 'pino';
 import withLogger from '../../withLogger';
 import { NOT_ACCEPTABLE, NOT_FOUND, NO_SUCCESS_RESPONSE_DEFINED } from '../errors';
@@ -26,11 +26,11 @@ import { ProblemJsonError } from '../../types';
 
 const outputNoContentFoundMessage = (contentTypes: string[]) => `Unable to find content for ${contentTypes}`;
 
-const findEmptyResponse = (response: IHttpOperationResponse, headers: IHttpHeaderParam[], mediaTypes: string[]) =>
+const createEmptyResponse = (code: string, headers: IHttpHeaderParam[], mediaTypes: string[]) =>
   pipe(
     mediaTypes,
     findIndex(ct => ct.includes('*/*')),
-    O.map(() => ({ code: response.code, headers }))
+    O.map(() => ({ code, headers }))
   );
 
 const helpers = {
@@ -101,7 +101,7 @@ const helpers = {
     const findHttpContent = hasContents(response)
       ? pipe(
           findDefaultContentType(response),
-          O.alt(() => findBestHttpContentByMediaType(response, ['application/json', '*/*']))
+          O.alt(() => findBestHttpContentByMediaType(response.contents, ['application/json', '*/*']))
         )
       : O.none;
 
@@ -138,7 +138,7 @@ const helpers = {
   },
 
   negotiateOptionsBySpecificResponse(
-    httpOperation: IHttpOperation,
+    requestMethod: string,
     desiredOptions: NegotiationOptions,
     response: IHttpOperationResponse
   ): RE.ReaderEither<Logger, Error, IHttpNegotiationResult> {
@@ -146,7 +146,7 @@ const helpers = {
     const { mediaTypes, dynamic, exampleKey } = desiredOptions;
 
     return logger => {
-      if (httpOperation.method === 'head') {
+      if (requestMethod === 'head') {
         logger.info(`Responding with an empty body to a HEAD request.`);
 
         return E.right({
@@ -155,61 +155,63 @@ const helpers = {
         });
       }
 
-      if (mediaTypes) {
-        // a user provided mediaType
-        const httpContent = hasContents(response) ? findBestHttpContentByMediaType(response, mediaTypes) : O.none;
-
-        return pipe(
-          httpContent,
-          O.fold(
-            () =>
-              pipe(
-                findEmptyResponse(response, headers || [], mediaTypes),
-                O.map(payloadlessResponse => {
-                  logger.info(`${outputNoContentFoundMessage(mediaTypes)}. Sending an empty response.`);
-
-                  return payloadlessResponse;
-                }),
-                E.fromOption<Error>(() => {
-                  logger.warn(outputNoContentFoundMessage(mediaTypes));
-
-                  return ProblemJsonError.fromTemplate(NOT_ACCEPTABLE, `Unable to find content for ${mediaTypes}`);
-                })
-              ),
-            content => {
-              logger.success(`Found a compatible content for ${mediaTypes}`);
-              // a httpContent for a provided mediaType exists
-              return pipe(
-                helpers.negotiateByPartialOptionsAndHttpContent(
-                  {
-                    code,
-                    dynamic,
-                    exampleKey,
-                  },
-                  content
-                ),
-                E.map(contentNegotiationResult => ({
-                  headers: headers || [],
-                  ...contentNegotiationResult,
-                  mediaType:
-                    contentNegotiationResult.mediaType === '*/*' ? 'text/plain' : contentNegotiationResult.mediaType,
-                }))
-              );
-            }
-          )
-        );
-      }
-      // user did not provide mediaType
-      // OR
-      // a httpContent for a provided mediaType does not exist
-      logger.trace('No mediaType provided. Fallbacking to the default media type (application/json)');
-      return helpers.negotiateDefaultMediaType(
-        {
-          code,
-          dynamic,
-          exampleKey,
-        },
-        response
+      return pipe(
+        O.fromNullable(mediaTypes),
+        O.chain(fromArray),
+        O.fold(
+          () => {
+            logger.debug('No mediaType provided. Fallbacking to the default media type (application/json)');
+            return helpers.negotiateDefaultMediaType(
+              {
+                code,
+                dynamic,
+                exampleKey,
+              },
+              response
+            );
+          },
+          mediaTypes =>
+            pipe(
+              O.fromNullable(response.contents),
+              O.chain(contents => findBestHttpContentByMediaType(contents, mediaTypes)),
+              O.fold(
+                () =>
+                  pipe(
+                    createEmptyResponse(response.code, headers || [], mediaTypes),
+                    O.map(payloadlessResponse => {
+                      logger.info(`${outputNoContentFoundMessage(mediaTypes)}. Sending an empty response.`);
+                      return payloadlessResponse;
+                    }),
+                    E.fromOption<Error>(() => {
+                      logger.warn(outputNoContentFoundMessage(mediaTypes));
+                      return ProblemJsonError.fromTemplate(NOT_ACCEPTABLE, `Unable to find content for ${mediaTypes}`);
+                    })
+                  ),
+                content => {
+                  logger.success(`Found a compatible content for ${mediaTypes}`);
+                  // a httpContent for a provided mediaType exists
+                  return pipe(
+                    helpers.negotiateByPartialOptionsAndHttpContent(
+                      {
+                        code,
+                        dynamic,
+                        exampleKey,
+                      },
+                      content
+                    ),
+                    E.map(contentNegotiationResult => ({
+                      headers: headers || [],
+                      ...contentNegotiationResult,
+                      mediaType:
+                        contentNegotiationResult.mediaType === '*/*'
+                          ? 'text/plain'
+                          : contentNegotiationResult.mediaType,
+                    }))
+                  );
+                }
+              )
+            )
+        )
       );
     };
   },
@@ -222,7 +224,7 @@ const helpers = {
       findLowest2xx(httpOperation.responses),
       RE.fromOption(() => ProblemJsonError.fromTemplate(NO_SUCCESS_RESPONSE_DEFINED)),
       RE.chain(lowest2xxResponse =>
-        helpers.negotiateOptionsBySpecificResponse(httpOperation, desiredOptions, lowest2xxResponse)
+        helpers.negotiateOptionsBySpecificResponse(httpOperation.method, desiredOptions, lowest2xxResponse)
       )
     );
   },
@@ -251,7 +253,7 @@ const helpers = {
           ),
           RE.chain(response =>
             pipe(
-              helpers.negotiateOptionsBySpecificResponse(httpOperation, desiredOptions, response),
+              helpers.negotiateOptionsBySpecificResponse(httpOperation.method, desiredOptions, response),
               RE.orElse(() =>
                 pipe(
                   helpers.negotiateOptionsForDefaultCode(httpOperation, desiredOptions),
@@ -289,19 +291,19 @@ const helpers = {
             pipe(
               previous,
               O.alt(() => {
-                logger.trace(`Unable to find a ${statusCodes[index]} response definition`);
+                logger.debug(`Unable to find a ${statusCodes[index]} response definition`);
                 return findResponseByStatusCode(httpResponses, current);
               })
             ),
           pipe(findResponseByStatusCode(httpResponses, first))
         ),
         O.alt(() => {
-          logger.trace(`Unable to find a ${tail(statusCodes)} response definition`);
+          logger.debug(`Unable to find a ${tail(statusCodes)} response definition`);
           return pipe(
             createResponseFromDefault(httpResponses, first),
             O.fold(
               () => {
-                logger.trace("Unable to find a 'default' response definition");
+                logger.debug("Unable to find a 'default' response definition");
                 return O.none;
               },
               response => {
@@ -341,7 +343,7 @@ const helpers = {
                 headers: response.headers || [],
               });
             } else {
-              logger.trace(`Unable to find a content with an example defined for the response ${response.code}`);
+              logger.debug(`Unable to find a content with an example defined for the response ${response.code}`);
               // find first response with a schema
               const responseWithSchema = response.contents && response.contents.find(content => !!content.schema);
               if (responseWithSchema) {
@@ -354,13 +356,9 @@ const helpers = {
                 });
               } else {
                 return pipe(
-                  findEmptyResponse(
-                    response,
-                    response.headers || [],
-                    get(contentWithExamples, 'mediaType', get(responseWithSchema, 'schema')) || ['*/*']
-                  ),
+                  createEmptyResponse(response.code, response.headers || [], ['*/*']),
                   E.fromOption(() => {
-                    logger.trace(`Unable to find a content with a schema defined for the response ${response.code}`);
+                    logger.debug(`Unable to find a content with a schema defined for the response ${response.code}`);
 
                     return new Error(`Neither schema nor example defined for ${response.code} response.`);
                   })